Some parameter in heat template should be string
Below parameters in kubecluster.yaml and kubeminion.yaml should be
string instead of boolean.

* flannel_use_vxlan
* kube_allow_priv
* rhn_register_host

For examples, "flannel_use_vxlan" is used in "write-flannel-config.sh", but
if this parameter is passed as boolean, it will be "True" as string not a
"true". As a result, "vxlan" is never enabled.

This patch fixes it.